* commit '08313ff69fb652296abb56a027b3c56f91f4eccd':
  SG-16213 dhcpcd: Fix hook-script to prevent periodic systemd-timesyncd restarts

+++ recipes-connectivity/dhcpcd/files/50-timesyncd.conf.sh
@@ -30,8 +30,10 @@ EOF
 
 remove_servers()
 {
-    rm -f "$SERVERFILE"
-    restart_needed=1
+    if [ -f "$SERVERFILE" ]; then
+        rm -f "$SERVERFILE"
+        restart_needed=1
+    fi
 }
 
 update_servers()
